Break out scroller component for reuse and augmentation with swipe directions and buttons
