+++ components/Scroller/Scroller.module.scss
@@ -1,0 +1,5 @@
+.horizontalScrollContainer {
+  margin: 1rem -1rem;
+  max-width: 100%;
+  max-height: max-content;
+}
