+++ components/Scroller/index.tsx
@@ -1,0 +1,3 @@
+import Scroller from "./Scroller"
+
+export default Scroller
