+++ components/index.tsx
@@ -4,6 +4,7 @@ import ImageWithFallback from "./ImageWithFallback"
 import Layout from "./Layout/Layout"
 import LoaderFlag from "./LoaderFlag"
 import Nav from "./Nav/Nav"
+import Scroller from "./Scroller"
 import Spacer from "./Spacer/Spacer"
 import {
   Avatar,
@@ -19,5 +20,6 @@ export {
   LoaderFlag,
   Nav,
   PartyAvatar,
+  Scroller,
   Spacer
 }
