+++ styles/politicianPage.module.scss
@@ -101,12 +101,6 @@
   }
 }
 
-.horizontalScrollContainer {
-  margin: 1rem -1rem;
-  max-width: 100%;
-  max-height: max-content;
-}
-
 .tagPage {
   display: flex;
   flex-wrap: wrap;
